requirements.lock.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
async-timeout==4.0.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
einops==0.8.1
et-xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.18.0
fonttools==4.58.2
frozenlist==1.6.2
fsspec==2025.5.1
ftfy==6.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gputil==1.4.0
greenlet==3.2.3
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
imagesize==1.4.1
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.2.17
langchain-community==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
lxml==5.4.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
mypy-extensions==1.1.0
networkx==3.4.2
numpy==1.24.4
openai==1.63.2
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.18
packaging==24.2
paddleocr==3.0.0
paddlepaddle==3.0.0
paddlex==3.0.0
pandas==1.5.3
pillow==11.2.1
pip==22.0.4
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.5
pydantic-core==2.33.2
pyparsing==3.2.3
pypdfium2==4.30.1
pyqt6==6.9.0
pyqt6-qt6==6.9.1
pyqt6-sip==13.10.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
ruamel-yaml==0.18.13
ruamel-yaml-clib==0.2.12
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
setuptools==58.1.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
sympy==1.14.0
tenacity==8.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
ujson==5.10.0
ultralytics==8.1.34
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
yarl==1.20.0
